+++ test/libsolidity/smtCheckerTests/functions/recursive_multi_return_2.sol
@@ -25,4 +25,3 @@ a;
 // ----
 // Warning: (72-90): Statement has no effect.
 // Warning: (96-107): Statement has no effect.
-// Warning: (304-307): Assertion checker does not support recursive function calls.