+++ test/libsolidity/smtCheckerTests/inheritance/constructor_state_variable_init_chain_run_all.sol
@@ -24,4 +24,10 @@ contract A is B {
 // ----
 // Warning: (171-176): Underflow (resulting value less than 0) happens here
 // Warning: (171-176): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (230-235): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (171-176): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (260-265): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (282-287): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (282-291): Overflow (resulting value larger than 2**256 - 1) happens here
+// Warning: (308-313): Overflow (resulting value larger than 2**256 - 1) happens here
 // Warning: (296-314): Assertion violation happens here