+++ test/libsolidity/smtCheckerTests/types/array_static_mapping_aliasing_2.sol
@@ -8,9 +8,9 @@ contract C
 	mapping (uint => uint)[2][2] severalMaps3d;
 	function f(mapping (uint => uint) storage map) internal {
 		map[0] = 42;
-		require(severalMaps[0][0] == 42);
-		require(severalMaps8[0][0] == 42);
-		require(severalMaps3d[0][0][0] == 42);
+		severalMaps[0][0] = 42;
+		severalMaps8[0][0] = 42;
+		severalMaps3d[0][0][0] = 42;
 		singleMap[0] = 2;
 		// Should not fail since singleMap == severalMaps[0] is not possible.
 		assert(severalMaps[0][0] == 42);
@@ -21,6 +21,10 @@ contract C
 		// Should fail since singleMap == map is possible.
 		assert(map[0] == 42);
 	}
+	function g(uint x) public {
+		f(severalMaps3d[x][0]);
+	}
 }
 // ----
-// Warning: (811-831): Assertion violation happens here
+// Warning: (781-801): Assertion violation happens here
+// Warning: (781-801): Assertion violation happens here