+++ test/libsolidity/smtCheckerTests/types/enum_in_struct.sol
@@ -1,19 +1,20 @@
 pragma experimental SMTChecker;
+pragma experimental ABIEncoderV2;
 
 contract C
 {
 	enum D { Left, Right }
 	struct S { uint x; D d; }
-	function f(S memory s) internal pure {
+	function f(S memory s) public pure {
 		s.d = D.Left;
 		assert(s.d == D.Left);
 	}
 }
 // ----
-// Warning: (109-119): Assertion checker does not yet support the type of this variable.
-// Warning: (139-142): Assertion checker does not yet support this expression.
-// Warning: (139-140): Assertion checker does not yet implement type struct C.S memory
-// Warning: (139-151): Assertion checker does not yet implement such assignments.
-// Warning: (162-165): Assertion checker does not yet support this expression.
-// Warning: (162-163): Assertion checker does not yet implement type struct C.S memory
-// Warning: (155-176): Assertion violation happens here
+// Warning: (143-153): Assertion checker does not yet support the type of this variable.
+// Warning: (171-174): Assertion checker does not yet support this expression.
+// Warning: (171-172): Assertion checker does not yet implement type struct C.S memory
+// Warning: (171-183): Assertion checker does not yet implement such assignments.
+// Warning: (194-197): Assertion checker does not yet support this expression.
+// Warning: (194-195): Assertion checker does not yet implement type struct C.S memory
+// Warning: (187-208): Assertion violation happens here